+++ NEWS.md
@@ -1,5 +1,7 @@
 # version 0.9-3
 
+* `st_is_valid` is now a generic
+
 * `plot.sf` gains an `extent` argument to set the extent (xlim, ylim) of the plot; `extent` must be an object with an `st_bbox()` method, such as an `sf` object; #1193
 
 # version 0.9-2